fix paths trigger for GH  deploy workflow
