+++ .github/workflows/deploy.yaml
@@ -1,12 +1,8 @@
-name: Deployment - Staging
+name: Deployment
 on:
   workflow_dispatch:
   push:
-    branches: [staging]
-    paths: ['src', 'prisma', 'Dockerfile-deploy', 'fly-staging.toml', '*.json']
-  pull_request:
-    branches: [staging]
-    paths: ['src', 'prisma', 'Dockerfile-deploy', 'fly-staging.toml', '*.json']
+    branches: [main, staging]
 
 env:
   FLY_API_TOKEN: ${{ secrets.FLY_API_TOKEN }}
@@ -44,4 +40,4 @@ jobs:
 
       - uses: superfly/flyctl-actions/setup-flyctl@master
       - name: Deploy
-        run: flyctl deploy --config fly-staging.toml --dockerfile Dockerfile-deploy
+        run: flyctl deploy --config fly-${{github.ref_name}}.toml --dockerfile Dockerfile-deploy
