+++ src/utils.ts
@@ -1,3 +1,5 @@
+import { Request, Response } from 'express'
+import { body, validationResult } from 'express-validator'
 import { CHUNK_SIZE } from '#'
 
 export const minusOneMonth = (date: Date) =>
@@ -23,3 +25,15 @@ export const filterSpaces =
   (minFollowers: number) =>
   ({ followers }: any) =>
     followers !== undefined && followers >= minFollowers
+
+const validations = [
+  body('maxOrgs').if(body('maxOrgs').exists()).isInt(),
+  body('minFollowers').if(body('minFollowers').exists()).isInt(),
+  body('since').if(body('since').exists()).isDate(),
+  body('until').if(body('until').exists()).isDate(),
+]
+
+export const validate = async (req: Request, res: Response) => {
+  await Promise.all(validations.map(async (validation) => validation.run(req)))
+  return validationResult(req)
+}
