+++ test/jest.config.ts
@@ -2,7 +2,14 @@ import type { JestConfigWithTsJest } from 'ts-jest'
 
 const jestConfig: JestConfigWithTsJest = {
   collectCoverage: true,
-  collectCoverageFrom: ['<rootDir>/src/**/*.ts'],
+  collectCoverageFrom: [
+    '<rootDir>/src/**/*.ts',
+    '!<rootDir>/src/service',
+    '!<rootDir>/src/snapshot',
+    '!<rootDir>/src/gh',
+    '!<rootDir>/src/daos',
+    '!<rootDir>/src/services/GroupService',
+  ],
   coverageDirectory: 'coverage',
   projects: [
     './test/jest.e2e.ts',
