+++ test/jest.lint.ts
@@ -7,7 +7,7 @@ const jestLintConfig: JestConfigWithTsJest = {
   displayName: 'lint',
   runner: 'jest-runner-eslint',
   testMatch: ['<rootDir>/**/*.(cjs|mjs|js|jsx|ts|tsx)'],
-  testPathIgnorePatterns: ['<rootDir>/test/coverage'],
+  testPathIgnorePatterns: ['<rootDir>/test/coverage', '<rootDir>/dist'],
 }
 
 export default jestLintConfig
