+++ test/jest.prettier.ts
@@ -30,6 +30,7 @@ const jestLintConfig: JestConfigWithTsJest = {
   testPathIgnorePatterns: [
     '<rootDir>/pnpm-lock.yaml',
     '<rootDir>/test/coverage',
+    '<rootDir>/dist',
     '<rootDir>/src/graphql/*.graphql',
     '<rootDir>/src/graphql/.graphqlconfig',
   ],
